pyproject.toml
[tool.poetry]
name = "sql-databricks-bridge"
version = "0.1.0"
description = "Bidirectional SQL Server ↔ Databricks bridge service"
authors = ["Joaquin Aldunate <joaquin.aldunate@wp.numerator.com>"]
readme = "README.md"
packages = [{include = "sql_databricks_bridge", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic-settings = "^2.1.0"
pyyaml = "^6.0.1"
pyjwt = "^2.8.0"
sqlalchemy = "^2.0.25"
pyodbc = "^5.1.0"
databricks-sdk = "^0.74.0"
polars = "^1.0.0"
pyarrow = "^15.0.0"
typer = "^0.12.0"
click = ">=8.0.0,<8.2.0"
rich = "^13.7.0"
watchfiles = "^0.21.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
httpx = "^0.27.0"
mypy = "^1.8.0"
ruff = "^0.2.0"

[tool.poetry.scripts]
sql-databricks-bridge = "sql_databricks_bridge.cli.commands:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
